+++ lib/flock.dart
@@ -2,7 +2,7 @@ library flock;
 
 export 'base/create_store.dart';
 export 'base/types.dart';
+export 'enhancers/async_publish.dart';
 export 'enhancers/store_proxy_base.dart';
-export 'enhancers/with_use_case.dart';
 export 'integration/project_to_stream.dart';
 export 'integration/store_builder.dart';